Changed id to bigint

+++ db/init/101_TBL_addresses.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS rpj.addresses (
-    id INT PRIMARY KEY,
-    street_id INT NOT NULL REFERENCES streets(id),
+    id BIGINT PRIMARY KEY,
+    street_id BIGINT NOT NULL REFERENCES streets(id),
     house_number VARCHAR(10) NOT NULL,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     geom GEOMETRY(POINT, 3765) NOT NULL
